Update AMDSFlatArray.cpp
still use quint32 for the size, and remove the warning

+++ source/DataElement/AMDSFlatArray.cpp
@@ -2,7 +2,7 @@
 
 #include <QDataStream>
 
-AMDSFlatArray::AMDSFlatArray(AMDSDataTypeDefinitions::DataType dataType, int size)
+AMDSFlatArray::AMDSFlatArray(AMDSDataTypeDefinitions::DataType dataType, quint32 size)
 {
 	resizeType(dataType, size);
 }
@@ -538,7 +538,7 @@ void AMDSFlatArray::clear(){
 	dataType_ = AMDSDataTypeDefinitions::InvalidType;
 }
 
-void AMDSFlatArray::clearAndReset(AMDSDataTypeDefinitions::DataType dataType, int size){
+void AMDSFlatArray::clearAndReset(AMDSDataTypeDefinitions::DataType dataType, quint32 size){
 	clear();
 	resizeType(dataType, size);
 }
@@ -717,8 +717,8 @@ bool AMDSFlatArray::resetTargetArrayAndReplaceData(AMDSFlatArray *targetArray) c
 	return copyDataToTargetArray(targetArray);
 }
 
-void AMDSFlatArray::resizeType(AMDSDataTypeDefinitions::DataType dataType, int size){
-	if( (dataType_ == dataType) && (size == this->size()) )
+void AMDSFlatArray::resizeType(AMDSDataTypeDefinitions::DataType dataType, quint32 size){
+	if( (dataType_ == dataType) && (size == (quint32)this->size()) )
 		return;
 
 	dataType_ = dataType;
